requirements.txt
fastapi
uvicorn
requests
streamlit
openai
python-dotenv
pytest
httpx
